requirements.txt
feedparser
openai
requests
beautifulsoup4
PyYAML
Jinja2
pydantic
